Tweaks to checksum.
Monotone-Parent: 84eee902dfd84cb65b81f79054d36ffe1470c786
Monotone-Revision: 3b9ce68a2339006abaf2088f3c4917a9b9c81407

+++ checksum.h
@@ -26,12 +26,10 @@
 #define CHECKSUM_H
 
 #include <stdlib.h>
+#include <stdint.h>
 
-/* This number must be prime, and must be smaller than 
- * (2^64)/(2^8) otherwise the multiplication may
- * overflow the >= 64-bit integer.
- */
-typedef unsigned long long checksum_integer_t;
+// We use the a value from Knuth's MMIX
+typedef uint64_t checksum_integer_t;
 static const checksum_integer_t LCG_A = 6364136223846793005ull;
 
 typedef struct
